+++ fractal/components/navigation-bars/NavigationBar/NavigationBar.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+    "title": "Navigation Bar",
+    "label": "Navigation Bar",
+    "order": 2,
+    "status": "ready",
+    "collated": true,
+    "preview": "@navigation-bar-preview",
+    "context": {
+        tabs: [
+            {label: "Lorem Ipsum A", to: "http://www.somewhere.com"},
+            {label: "Lorem Ipsum B", to: "http://www.elsewhere.com", selected: true},
+            {label: "Lorem Ipsum C", to: "http://www.everywhere.com"}
+        ]
+    }
+};