+++ fractal/components/navigation-bars/NavigationTab/NavigationTab.config.js
@@ -1,0 +1,17 @@
+module.exports = {
+    "title": "Navigation Bar Linked Tab",
+    "label": "Navigation Tab",
+    "order": 1,
+    "status": "ready",
+    "collated": true,
+    "preview": "@navigation-bar-preview",
+    "context": {
+        "label": "Lorem Ipsum",
+        "to": "http://www.somewhere.com"
+    },
+    "variants": [
+        {"name": "navigation-tab-selected", "context": {
+            "selected": true
+        }}
+    ]
+};