+++ fractal/components/navigation-bars/SecondaryNavigationBar/SecondaryNavigationBar.config.js
@@ -1,0 +1,13 @@
+const navigationConfig = require('../NavigationBar/NavigationBar.config');
+
+module.exports = {
+    "title": "Secondary Navigation Bar",
+    "label": "Secondary Navigation Bar",
+    "order": 4,
+    "status": "ready",
+    "collated": true,
+    "preview": "@navigation-bar-preview",
+    "context": {
+        tabs: navigationConfig.context.tabs
+    }
+};